RavenDB-7070 moving very slow tests to stress
