+++ test/StressTests/Server/Documents/PeriodicBackup/PeriodicBackupTestsStress.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Newtonsoft.Json;
 using Raven.Client.Documents.Operations;
 using Raven.Client.Documents.Operations.Backups;
 using Raven.Client.ServerWide.Operations;
@@ -90,9 +91,9 @@ namespace StressTests.Server.Documents.PeriodicBackup
             var backupPath = NewDataPath(suffix: "BackupFolder");
             using (var server = GetNewServer())
             using (var store = GetDocumentStore(new Options
-            {
-                Server = server
-            }))
+                   {
+                       Server = server
+                   }))
             {
                 using (var session = store.OpenAsyncSession())
                 {
@@ -147,7 +148,7 @@ namespace StressTests.Server.Documents.PeriodicBackup
             (List<RavenServer> nodes, RavenServer leaderServer) = await CreateRaftCluster(clusterSize);
             var mentorNode = nodes.First(x => x != leaderServer);
 
-            using (var store = GetDocumentStore(new Options { Server = leaderServer, ReplicationFactor = 3}))
+            using (var store = GetDocumentStore(new Options { Server = leaderServer, ReplicationFactor = 3 }))
             {
                 await Backup.FillClusterDatabaseWithRandomDataAsync(databaseSizeInMb: 1, store, clusterSize);
 
@@ -218,14 +219,14 @@ namespace StressTests.Server.Documents.PeriodicBackup
                 await Backup.HoldBackupExecutionIfNeededAndInvoke(database.PeriodicBackupRunner.ForTestingPurposesOnly(), async () =>
                 {
                     WaitForValue(() =>
-                    {
-                        var now = DateTime.Now;
-                        return now.Minute % 2 == 0 && now.Second <= 10;
-                    },
-                       expectedVal: true,
-                       timeout: (int)TimeSpan.FromMinutes(2).TotalMilliseconds,
-                       interval: (int)TimeSpan.FromSeconds(1).TotalMilliseconds
-                   );
+                        {
+                            var now = DateTime.Now;
+                            return now.Minute % 2 == 0 && now.Second <= 10;
+                        },
+                        expectedVal: true,
+                        timeout: (int)TimeSpan.FromMinutes(2).TotalMilliseconds,
+                        interval: (int)TimeSpan.FromSeconds(1).TotalMilliseconds
+                    );
 
                     var config = Backup.CreateBackupConfiguration(backupPath, fullBackupFrequency: fullBackupFrequency);
                     var taskId = await Backup.UpdateConfigAndRunBackupAsync(server, config, store, opStatus: OperationStatus.InProgress);
@@ -255,10 +256,48 @@ namespace StressTests.Server.Documents.PeriodicBackup
 
                     Assert.Equal(backupStatus.OriginalBackupTime,
                         delayUntil < nextFullBackup
-                            ? taskBackupInfo.OnGoingBackup.StartTime    // until the next scheduled backup time.
-                            : nextFullBackup.Value.ToUniversalTime());  // after the next scheduled backup.
+                            ? taskBackupInfo.OnGoingBackup.StartTime // until the next scheduled backup time.
+                            : nextFullBackup.Value.ToUniversalTime()); // after the next scheduled backup.
                 }, tcs: new TaskCompletionSource<object>(TaskCreationOptions.RunContinuationsAsynchronously));
             }
+        }
+
+        [RavenFact(RavenTestCategory.BackupExportImport)]
+        public async Task NextCronScheduleOccurence_BasedOnLastBackup_ShouldBeCorrect()
+        {
+            const string endpoint = "/studio-tasks/next-cron-expression-occurrence";
+            const string cronExpression = "*/2 * * * *";
+            using var store = GetDocumentStore();
+
+            var configuration = Backup.CreateBackupConfiguration(NewDataPath(), fullBackupFrequency: cronExpression);
+            await Backup.WaitUntilNextFullBackupActionWindowAsync(configuration, TimeSpan.FromSeconds(15), Server.ServerStore.ServerShutdown);
+            var taskId = await Backup.UpdateConfigAsync(Server, configuration, store);
+            await Task.Delay(TimeSpan.FromSeconds(130));
+
+            var client = store.GetRequestExecutor().HttpClient;
+
+            var uri = $"{store.Urls.First()}{endpoint}?expression=* * * * *&taskId={taskId}&database={store.Database}&isFull=true";
+            var json = await client.GetStringAsync(uri);
+            var response = JsonConvert.DeserializeObject<NextCronExpressionOccurrenceResponse>(json);
+
+            // The endpoint with taskId should return the next occurrence based on the last backup time
+            Assert.True(response.IsValid, $"Expected valid response, but got: {json}");
+            Assert.True(response.Utc < DateTime.UtcNow, $"Based on the last backup time, the next cron schedule occurrence should be in the past, but got UTC: {response.Utc}, ServerTime: {response.ServerTime}");
+
+            uri = $"{store.Urls.First()}{endpoint}?expression=* * * * *";
+            json = await client.GetStringAsync(uri);
+            response = JsonConvert.DeserializeObject<NextCronExpressionOccurrenceResponse>(json);
+
+            // The endpoint without taskId should return the next occurrence based on the current time
+            Assert.True(response.IsValid, $"Expected valid response, but got: {json}");
+            Assert.True(response.Utc > DateTime.UtcNow, $"Expected next cron schedule occurrence to be in the future, but got UTC: {response.Utc}, ServerTime: {response.ServerTime}");
+        }
+
+        private record NextCronExpressionOccurrenceResponse
+        {
+            public bool IsValid { get; init; }
+            public DateTime Utc { get; init; }
+            public DateTime ServerTime { get; init; }
         }
     }
 }
